refactor!: rename log feature to tracing-log
Previously `quinn*` would provide the `log` feature to log events via `log` if
no `tracing` subscriber exists.

Later https://github.com/quinn-rs/quinn/pull/1923 allowed `quinn-udp` to log via
`log` directly, making `tracing` an optional dependency. For that, it introduced
the `direct-log` feature, a workaround name in order to not introduce a breaking
change.

This commit cleans up the above, renaming the `log` feature to `tracing-log` and
the `direct-log` to `log`. This is a breaking change and thus `quinn-udp` is
bumped to `v0.6.0`.

See https://github.com/quinn-rs/quinn/pull/1921 for the full history.